리팩토링: 사용자 엔티티 중복코드 조회
- 이번에는 테스트가 있어서 좋았다.
